Relocate fetchService
