+++ src/api/AuthRoutes.ts
@@ -1,5 +1,5 @@
 import Config from '../config';
-import { fetchService } from '../utils';
+import fetchService from './fetchService';
 import { LoginRequest, RegistrationRequest, PasswordResetRequest, EmailModificationRequest } from './ApiRequests';
 import { RegistrationResponse, LoginResponse, VerifyAuthTokenResponse } from './ApiResponses';
